Split change_ and fix_ visibility assists
